Reorganize repository
Reorganize the repository so that only polybar-4 is in the repository,
remove most of the README, etc. That is, strip it down to what I'm
using so I can just clone this into ~/.config/polybar and not worry
about the details.

There are a bunch of modifications to the polybar-4 configs, some of
them likely don't work, etc.