Update polybar config somewhat
  * Update colors and make foreground / background env configurable.
    This is to match equivalent changes to the herbstluftwm config,
    which sources a conf file when available to set most of the config
    variables used here.

  * Add various places for configuration via environment variable, such
    as modules, network interfaces, batteries, whether the bar goes on
    the top or bottom of the screen, and so on.

  * Remove unused or unwanted modules.

  * Rename wireless-network and wired-network to wlan and lan,
    respectively.

  * Adjust width and margins.

  * Remove the separator string.

  * Clean up file headings, EOF comments, and separator comments. These
    were unusually gaudy and I really don't need a giant "BARS" every
    time I open a file. It just needs to be one line.

  * Delete bars.ini and user_modules.ini entirely. These files have no
    use in my config, so might as well remove them.

A lot of this is to support some changes to my herbstluftwm config and
to allow local config, since each of my machines has a different
PulseAudio sink, different WLAN and LAN interfaces (some just don't have
LAN), and so on. At least one machine has slightly larger font sizes to
account for DPI scaling on Linux being a very mixed bag, and so on and
so forth.

A conf.example is provide that should be named 'conf' for the
herbstluftwm config, but I already know what I'm doing there.